Remount with the noatime option remove.
Remounting can be dangerous at times.  I believe I heard this has to do
with if a user has set file system type to "auto" in their `/etc/fstab`.
This configuration is also probably unnecessary because the kernel sets
"relatime" as the default option now (I heard this, and it does on
mine).